Bug 1238522 - Don't use float type for absmiddle attirubte. r=eeejay
SAPI's absmiddle attibute doesn't allow float type.  Example, when pitch was 0.75, abmiddle was -2.5.  So ISpVoice.Speak() will return error.

MozReview-Commit-ID: 2AnrSEDTFkl

UltraBlame original commit: d5f3cacd2c7523dee9a19e1853336fce9094180a

+++ dom/media/webspeech/synth/windows/SapiService.cpp
@@ -369,7 +369,8 @@ SapiService::Speak(const nsAString& aText, const nsAString& aUri,
   
   nsAutoString xml;
   xml.AssignLiteral("<pitch absmiddle=\"");
-  xml.AppendFloat(aPitch * 10.0f - 10.0f);
+  
+  xml.AppendInt(static_cast<int32_t>(aPitch * 10.0f - 10.0f));
   xml.AppendLiteral("\">");
   uint32_t textOffset = xml.Length();
